Silent aim option and hitbox deleted
Unfortunately, the silent option doesn't work against players, only against NPCs, so I've decided to remove it from PvP mode since it doesn't really do anything. That said, it can still be used just fine in PvE mode against NPCs. I thought it worked, but it actually doesn't. When I first tested it, it was in a public match and I didn't realize it wasn't working. But after some tests in a private match, I came to the conclusion that the hitbox for players only expands visually on the client side — unlike the old method, which did work but has now been detected, as I mentioned in previous updates.